Task11

+++ restaurants/views.py
@@ -3,6 +3,9 @@ from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
 
+def no_access(request):
+    return render(request, 'no_access.html')
+
 def signup(request):
     form = SignupForm()
     if request.method == 'POST':
@@ -59,6 +62,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -73,8 +78,11 @@ def restaurant_create(request):
     return render(request, 'create.html', context)
 
 def item_create(request, restaurant_id):
+   
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user == restaurant.owner or request.user.is_staff):
+        return redirect('no_access' )
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -89,7 +97,10 @@ def item_create(request, restaurant_id):
     return render(request, 'item_create.html', context)
 
 def restaurant_update(request, restaurant_id):
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user == restaurant_obj.owner or  request.user.is_staff):
+        return redirect('no_access')
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -103,6 +114,8 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+    if not request.user.is_staff:
+        return redirect('no_access')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
